feat: debug-materials (#569)
* fix: require vscode (#565)

* feat: generate-debug-material-json

* refactor: record (#568)

* refactor: record execute command

* style: remove empty line

* feat: record for service

* chore: rename module

* feat: record open in browser

* feat: record show doc pick

* refactor: no need locale

* refactor: remove data

* chore: remove console

* feat: update

* feat: update

* feat: update readme

* fix: rules

* feat: support vue (#580)

* feat: support scan command (#581)

* chore: sets the extension to be flagged as a Preview in the Marketplace. (#582)

* feat: code Editor for Page Templater (#583)

* feat: add editor

* docs: changlog

* fix: no select blocks prompt (#575)

* fix: select blocks prompt

* chore: version and changelog

* fix: open materials panel when code editor opened (#573)

* feat: unable to open material panel when no editor

* feat: input path in settings.json

* feat: support generate page or component to the specify path

* fix: configuration description i18n

* fix: lint

* feat: support generate code to special path

* feat: generate code to special path

* feat: add debug function to bulk download

* feat: add delete debug material source function

* chore: i18n

* chore: move files and add deps

* feat: custom material sources will be at first in material sources list (#586)

* chore: update version and change log

* docs: new image for Visual Construction

* fix: require path not found (#590)

* chore: sorting of quick entry (#589)

* chore: sorting of quick entry

* docs: typo

* chore: package version

* feat: update readme

* chore: move debug material more quickly

* feat: choose material folder

* chore: title for quick entry (#592)

* docs: update docs and versions

* fix: project packageJSON dependencies not found

* fix: block display default page

* fix: screenshot

* fix: import image-to-base64

* chore: material-ui version

* chore: image to base64 package version

* fix: comment

* fix: rename

+++ packages/common-service/src/index.ts
@@ -374,7 +374,7 @@ export const bulkInstallMaterialsDependencies = async function (
   // filter existing dependencies of project
   const filterDependencies: { [packageName: string]: string }[] = [];
   Object.keys(pagesDependencies).forEach((packageName) => {
-    if (!projectPackageJSON.dependencies.hasOwnProperty(packageName)) {
+    if (projectPackageJSON.dependencies && !projectPackageJSON.dependencies.hasOwnProperty(packageName)) {
       filterDependencies.push({
         [packageName]: pagesDependencies[packageName],
       });
@@ -410,31 +410,43 @@ export const bulkDownloadMaterials = async function (
     materials.map(async (material: any) => {
       await fse.mkdirp(tmpPath);
       const materialName: string = upperCamelCase(material.name);
-
-      let tarballURL: string;
-      try {
-        log(i18n.format('package.common-service.downloadMaterial.getDownloadUrl'));
-        tarballURL = await getTarballURLByMaterielSource(material.source);
-      } catch (error) {
-        error.message = i18n.format('package.common-service.downloadMaterial.downloadError', {
-          materialName,
-          tarballURL,
-        });
-        throw error;
-      }
-      log(i18n.format('package.common-service.downloadMaterial.unzipCode'));
       const downloadPath = path.join(tmpPath, materialName);
-      try {
-        await getAndExtractTarball(downloadPath, tarballURL, ({ percent }) => {
-          log(i18n.format('package.common-service.downloadMaterial.process', { percent: (percent * 100).toFixed(2) }));
-        });
-      } catch (error) {
-        error.message = i18n.format('package.common-service.uzipError', { materialName, tarballURL });
-        if (error.code === 'ETIMEDOUT' || error.code === 'ESOCKETTIMEDOUT') {
-          error.message = i18n.format('package.common-service.uzipOutTime', { materialName, tarballURL });
+      // debug local material application
+      if (material.source.type === 'debug') {
+        try {
+          await fse.copy(material.source.path, downloadPath, {
+            filter: srcPath => {
+              return !srcPath.includes('node_modules');
+            }
+          });
+        } catch (err) {
+          log(i18n.format('package.common-service.downloadMaterial.debugDownloadError', { errMessage: err.message }))
         }
-        await fse.remove(tmpPath);
-        throw error;
+      } else {
+        let tarballURL: string;
+        try {
+          log(i18n.format('package.common-service.downloadMaterial.getDownloadUrl'));
+          tarballURL = await getTarballURLByMaterielSource(material.source);
+        } catch (error) {
+          error.message = i18n.format('package.common-service.downloadMaterial.downloadError', {
+            materialName,
+            tarballURL,
+          });
+          throw error;
+        }
+        log(i18n.format('package.common-service.downloadMaterial.unzipCode'));
+        try {
+          await getAndExtractTarball(downloadPath, tarballURL, ({ percent }) => {
+            log(i18n.format('package.common-service.downloadMaterial.process', { percent: (percent * 100).toFixed(2) }));
+          });
+        } catch (error) {
+          error.message = i18n.format('package.common-service.uzipError', { materialName, tarballURL });
+          if (error.code === 'ETIMEDOUT' || error.code === 'ESOCKETTIMEDOUT') {
+            error.message = i18n.format('package.common-service.uzipOutTime', { materialName, tarballURL });
+          }
+          await fse.remove(tmpPath);
+          throw error;
+        }
       }
     })
   );
